made a LandingPage component that renders the PrimaryPage component when the get-started button is clicked using state
