+++ src/App.js
@@ -1,82 +1,22 @@
-import { useState, useEffect, useReducer } from "react";
-
+import React, { useState } from "react";
 import "./App.css";
-import { Cards } from "./components/Cards";
-import { SelectionInfo } from "./components/SelectionInfo/SelectionInfo";
-import { SelectedMembers } from "./components/SelectedMembers/SelectedMembers";
-import { ErrorMessage } from "./components/ErrorMessage/ErrorMessage";
-import { PromptMessage } from "./components/PromptMessage/PromptMessage";
-import { teamSelectionReducer } from "./state/reducers/teamSelectionReducer";
-
-const initialState = {
-  allUsers: [],
-  selectedUser: "",
-  chosenTeamMembers: [],
-  errorMessage: "",
-  promptMessage: "Please Select Users",
-};
+import { PrimaryPage } from "./pages/PrimaryPage/PrimaryPage";
+import { LandingPage } from "./pages/LandingPage/LandingPage";
 
 function App() {
-  const [state, dispatch] = useReducer(teamSelectionReducer, initialState);
+  const [showPrimaryPage, setShowPrimaryPage] = useState(false);
 
-  useEffect(() => {
-    fetch("https://jsonplaceholder.typicode.com/users")
-      .then(response => response.json())
-      // payload is users instead of []
-      .then(users => dispatch({ type: "SET_USERS", payload: users }));
-  }, []);
-
-  function handleOnCLick() {
-    // payload is switch function instead of [] for setTeamMembers & currentState is now state.teamMembers
-
-    switch (true) {
-      case state.chosenTeamMembers.length > 2:
-        dispatch({
-          type: "SET_ERROR_MESSAGE",
-          payload: "Maximum Limit Reached",
-        });
-
-        break;
-      case state.selectedUser === "":
-        dispatch({
-          type: "SET_ERROR_MESSAGE",
-          payload: "No User Selected",
-        });
-        break;
-
-      case state.chosenTeamMembers.includes(state.selectedUser):
-        dispatch({
-          type: "SET_ERROR_MESSAGE",
-          payload: "Teammember Already Selected",
-        });
-        break;
-      default:
-        dispatch({
-          type: "SET_TEAM_MEMBERS",
-          payload: state.selectedUser,
-        });
-    }
-  }
+  const handleButtonClick = () => {
+    setShowPrimaryPage(true);
+  };
 
   return (
-    <div className="App">
-      <div className="body">
-        <div className="left-column">
-          <SelectionInfo userName={state.selectedUser} />
-          <button onClick={handleOnCLick}>Add Teammate</button>
-          <SelectedMembers chosenTeamMembers={state.chosenTeamMembers} />
-          <ErrorMessage errorMessage={state.errorMessage} />
-          <PromptMessage promptMessage={state.promptMessage} />
-        </div>
-        <div className="right-column">
-          <Cards
-            allUsers={state.allUsers}
-            selectedUsers={state.selectedUsers}
-            dispatch={dispatch}
-            chosenTeamMembers={state.chosenTeamMembers}
-          />
-        </div>
-      </div>
+    <div>
+      {showPrimaryPage ? (
+        <PrimaryPage />
+      ) : (
+        <LandingPage onButtonClick={handleButtonClick} />
+      )}
     </div>
   );
 }
